+++ src/vis/bar/utils/config-util.ts
