+++ src/vis/bar/utils/index.ts
@@ -1,1 +1,2 @@
-export * from './utils';
+export * from './config-util';
+export * from './number-formatter';
