+++ src/vis/bar/utils/number-formatter.ts
@@ -1,0 +1,6 @@
+export const numberFormatter = new Intl.NumberFormat('en-US', {
+  maximumFractionDigits: 4,
+  maximumSignificantDigits: 4,
+  notation: 'compact',
+  compactDisplay: 'short',
+});
